use scroll on init in PageTreeSubstance
